OPTIMUS-36426 export changes from b942b22badb3640f7c87162fd4fce7dfa3f2a6a1
original_id:b492bd1124b828566cefb21ac4a105bf19add27c

+++ optimus/platform/projects/core/src/main/scala/optimus/graph/ProfilingEventStatsFileUtil.scala
@@ -41,7 +41,7 @@ object ProfilingEventStatsFileUtil {
 
   private def convertJsValue(value: JsValue): Any = {
     value match {
-      case JsNumber(n) => n.intValue
+      case JsNumber(n) => n.doubleValue
       case JsString(n) => n
       case _           => value
     }
